refactor the tests directory to break down by files, so imports are cleaner, and it's easier to find which tests i'm looking at
