+++ tests/images/README.txt
@@ -1,0 +1,5 @@
+This folder contains images used in tests.
+
+More may be added at a later time to test different file formats.
+
+Try to keep the files small (find or create smaller files!)... exiftool is for metdata manipulation, not image data!
