+++ tests/images/skyblue.png
